Add ft_bzero.c & test_ft_bzero.c files

+++ ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: aborda <aborda@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/05 09:41:32 by aborda            #+#    #+#             */
-/*   Updated: 2025/08/05 10:44:39 by aborda           ###   ########.fr       */
+/*   Updated: 2025/08/05 20:55:39 by aborda           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memset(void *s, int c, size_t n)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
